Add error descriptions to header info panel
Summary:
This changeset adds useful descriptions to the violations reported in the
info panel by the header plugin.

I also intelligently keep the info panel visible if two plugins are active
at once where one of the plugins does not write to the panel. In the future
I will only allow one plugin to be activated at a time, so this logic will
be removed.

Test Plan:
. `webpack-dev-server --hot`
. `locahost:8080/webpack-dev-server`
. Activate headers plugin
. See the panel contains a violation on an `<h3>`
. Hover over the red `3` label to see a description of the error

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D17837