Added a plugin to detect unlabeled inputs
Summary:
This plugin highlights inputs with missing labels. Reporting errors is
currently limited to highlighting the input and displaying a tooltip on hover.
In the future I'd like to revisit this plugin and populate something in the
info panel, but as I am still building this toolbar out I will revisit it.

Some design decisions here are fairly opinionated, but I followed the
guidelines in http://webaim.org/techniques/forms/controls as a reference.

Here's a gif of it in action:

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Activate the Labels plugin
. See some text boxes highlighted in red. One box does not have a label,
another is wrapped in a label tag with no text content. Another is a button
with no text.

Reviewers: michelle, john

Subscribers: riley

Differential Revision: https://phabricator.khanacademy.org/D17895