Fixed placement of labels and highlights
Summary:
In this changeset I've embraced `.position()` and `.offsetParent()`, as well
as `.outerWidth()` and `.outerHeight()`.

I brought in a bootstrap page for testing and found that I was highlighting
incorrectly for fixed elements. Attempting to fix it, I broke highlighting
for everything else. Then I realized I was using jQuery and didn't have to
put up with this madness.

I also phased out `.highlightText()`.

Test Plan:
. `webpack-dev-server --hot`
. `http://localhost:8080/webpack-dev-server/`
. Fingers crossed for this one, only because I have not changed index.html
. But index.html should behave exactly the same

Reviewers: john, michelle

Differential Revision: https://phabricator.khanacademy.org/D17907