Added a plugin to mark ARIA landmarks
Summary:
Simple plugin that scans for anything with a role and displays it in a label.

I also replaced `index.html` with a bootstrap template, which I'll be slowly
modifying to include information on dinosaurs. This helped me discover some
issues with my highlight code that has been fixed in D17907.

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server/`
. Check ou

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D17909