Added namespacing for annotations
Summary:
This changeset adds namespaces to the annotation utility. Plugins can include the annotation
module, and invoke the now-returned function with a string. That string will be used to
append an extra class to any annotations, and allow the plugins to simply call
`annotate.removeAll()` to remove them.

As such, I've removed the `className` argument to both `label` and `highlight`. Both of
these functions return the annotation as a jQuery object, so the callee can simply
call `addClass()` on the returned value (this is used in the labels plugin).

Test Plan:
. No visual changes!
. `webpack-dev-server --hot`
. `locahost:8080/webpack-dev-server/`
. Verify that the plugins work correctly

Reviewers: michelle, john

Subscribers: riley

Differential Revision: https://phabricator.khanacademy.org/D18004