Refactored out the info panel to its own shared component
Summary:
We used to render the info panel based on html returned from each of the
plugins. This is going to be very difficult to maintain if we want two
plugins to display information in the panel simultaneously.

I've moved the logic for rendering the panel into its own class, which plugins
can call directly. This means that info panels are now unique to the plugin
which created them.

I've also added tabs to the info panel. Currently there is an "About" tab, and
a "Summary." In a future diff I will be adding an "Errors" tab - a normalized
way to display errors from a plugin.

These panels will also eventually be moveable and collapseable. Not yet, so
I can keep this diff small.

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Activate the "Headers" plugin
. See an info panel appear with the "Summary" tab active
. See you can switch between "About" and "Summary"

Reviewers: michelle

Subscribers: riley

Differential Revision: https://phabricator.khanacademy.org/D18110