Changed behavior to only allow one plugin to be active at a time
Summary:
I had a revision like this recently but abandoned it after deciding that we
should allow multiple plugins to be active.

I've changed my mind again, this time for presentation reasons. It's going to
be very noisy displaying a summary and errors for multiple plugins, and there
is no real benefit to allowing simultaneous plugins.

As such I've added a custom UI element that's a mix between a radio button and
a checkbox. This is because checkboxes allow multiple items to be selected,
and radios do not allow users to de-select. Consider this a hybrid.

Here's a GIF of it in action:

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Click "Headers"
. See the checkbox
. Click "Labels"
. See that headers is now unchecked
. Click "Labels"
. See that nothing is checked

Reviewers: riley, michelle

Differential Revision: https://phabricator.khanacademy.org/D18146