Add basic error reporting and interactions (WIP)
Summary:
This changeset adds a basic error tab to the info panel, which allows plugins
to report errors and display them in a controlled way. The `addError` method
also accepts a referenced jQuery element, which will be highlighted when the
user hovers over the error.

This is very much a WIP for easy reviewing. In a future diff I'll be adding
useful error messages to the Headers plugin and further refining the design
of this new tab.

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Activate the headers plugin
. See the new "Errors" tab is now active
. Hover over the first error and see the invalid <h3> is highlighted
. See a count of one next to the tab

Reviewers: michelle, riley

Differential Revision: https://phabricator.khanacademy.org/D18149