Made the info panels draggable
Summary:
This changeset allows users to move the info panel by dragging its title bar,
revealing elements behind it that may be causing violations and appearing in
the info panel itself.

I also made it so that highlight annotations appear above the panel, and
refactored some z-index values while I was in the neighborhood.

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Activate the headings plugin
. Mouse over the title bar of the info panel
. See that you can drag the info panel by its title bar
. See that the "X" button still works

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D18222