Allow labels to provide extended descriptions on hover
Summary:
This changeset helps users get a glimpse of accessibility errors without the
need to reference the info panel.

There is now a special `errorLabel` function in the annotate module which
creates a label with an extended description that appears on hover. I'll be
using this for alt-text as well in a future diff.

I'm hoping that this somewhat solves the "below the fold" problem, since users
will no longer need the info panel to tell them the errors on the page. I'll
keep iterating on the idea.

Test Plan:
. `webpack-dev-server`
. `localhost:8080/webpack-dev-server`
. Activate the headings plugin
. See headings with red labels
. Hover over them and see an extended description

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D18243