Added a plugin to label text contrast and suggest appropriate alternatives
Summary:
This changeset adds a contrast plugin, which makes extensive use of Chrome's
Accessibility Developer Tools that I brought in for D18211.

It labels items with (unique) text contrast (as to not bombard the screen with
50 identical labels). If the contrast is good enough, the label is green. If
not, the label is red and contains a viable alternative on hover.

No use of the info panel yet - that's coming soon.

Test Plan:
. `webpack-dev-server --hot`
. `localhost:8080/webpack-dev-server`
. Activate the "Contrast" plugin
. See a few green labels, and a single red label on the Sign up button
. Hover over the signup button and see a suggested background color

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D18270