Fixed logic for detecting "invisible" elements
Summary:
We were using `axs.utils.elementIsVisible`, which checks a combination of
things including:

- If the element has zero area
- If the element is invisible
- If the element is behind other elements

We do not need the third bullet point, so I've rewritten this line to only
omit invisible elements or elements with zero area from contrast testing.

This popped up because the list of plugins is getting quite long as often
covers elements that it aims to test for.

Test Plan:
. Activate Contrast plugin
. See the first link, and the "See more" button are detected properly

Reviewers: michelle

Differential Revision: https://phabricator.khanacademy.org/D18301