Queue up label annotations to reduce reflows
Summary:
While running this toolbar on more complicated interfaces, the time to create
labels is incredible slow and, at times, crashes the page.

To reduce reflows, I've changed the behavior of the `annotate.label` method
to queue up annotations, which are then mounted to the DOM on every frame.

Renders occur in chunks for better performance. The constant `RENDER_CHUNK`
currently has a value of 20, which through testing offers 60fps performance
on my facebook homepage. We could likely double or triple this on sites with
less markup (recalc costs are lower), but on smaller sites we're less likely
to make that many annotations in the first place.

Test Plan:
. `webpack && cat dist/tota11y.js| pbcopy`
. Head over to FB.com
. Paste into the console
. Activate the "Contrast" plugin
. See the page doesn't hang as the labels pour in

Reviewers: riley

Differential Revision: https://phabricator.khanacademy.org/D18386