Establish mapping between error labels and entries in the info panel
Summary:
It's currently possible to locate errors from the info panel by using the
small magnifying glass icon. The opposite direction, however, is quite
difficult.

This changeset adds a small link to error labels that, when clicked, brings
up the expanded error message in the info panel. If the info panel is closed,
it will automatically reopen.

(gif)

Test Plan:
. Activate headers plugin
. Scroll down to last error
. Click the [?] in the error label
. See the error is expanded in the info panel
. Close the info panel
. Click the [?] icon
. See the info panel reopens

Reviewers: riley

Differential Revision: https://phabricator.khanacademy.org/D18414