Highlight info panel errors from error labels
Summary:
In D18414 we add functionality for users to navigate to info panel errors
from error labels, but there is very little feedback. To further build out
two-way navigate between the info panel and labels, we now highlight the
entries in the info panel when hovering over the error labels.

(gif)

Test Plan:
. Activate Headings plugin
. Navigate to the last error
. Hover over the error label (the one with h4)
. See the corresponding info panel entry is highlighted

Reviewers: riley

Differential Revision: https://phabricator.khanacademy.org/D18432