Added a bookmarklet to installation section
Summary:
This changeset adds a bookmarklet that users can drag into their bookmarks
bar. It's hidden on mobile.

I was originally going to prevent this AJAX from occuring on mobile, but
then tried to deal with users who may load the screen at a small width and
then expand on desktop. It become a real hassle, and then I realized this
request is most likely cached.

Test Plan:
. Open up the landing page
. See a bookmarklet under the "Installation" section
. See that you are able to drag it into your bookmarks bar
. See that clicking this (on another page) inserts a full-function tota11y instance into the page

Reviewers: riley

Differential Revision: https://phabricator.khanacademy.org/D18485