Fixed bookmarklet button
Ref #9, #10

This button was populated by an AJAX request, which inlines the `href` with
the contents of `tota11y.min.js`.

After changing the location of `tota11y.min.js` in 0816231, I neglected to
update this code. This is a quick fix to get the button working again, but
we'll likely want to explore a long-term solution of having the bookmarklet
create a script tag, which serves `tota11y.min.js` from a CDN.