Added mock DOM and several stubs to allow plugin testing with mocha
Summary:
Using jsdom, we can now create an environment in which to unit test tota11y
plugins. Currently, this is limited to the Landmarks plugin, which applies labels
to elements with a role attribute.

We accomplish this by changing the behavior of `require` to do the following:
* Ignore less/handlebars loaders (i.e. `require("./template.handlebars");`)
* Provide our own instance of jQuery (to make the plugins work in the jsdom environment)
* Stub out the annotations module (to allow us to query for annotations in the new environment)

Tests are written in ES6 as well, using a special flag to `mocha`.

This is a first step towards providing solid unit tests for our plugins.

Test Plan: . npm test

Reviewers: riley

Differential Revision: https://phabricator.khanacademy.org/D18877