Contrast plugin suggestion preview - Fixes #21
This adds a checkbox next to the suggested color combination which will toggle between original colors and suggested colors. In order to embed the functionality for the checkbox I had to modify all plugins to pass their descriptions as jQuery objects (with the blessing of @jdan of course).

It might be worth discussing if there is a better way to represent the control for this than a checkbox; it seemed like the cleanest and simplest way to me but I can see that it might not be very intuitive, especially without a label.