Patch axs.AuditRule.collectMatchingElements to prevent JS errors with cross-origin iframes
Summary:
https://github.com/GoogleChrome/accessibility-developer-tools/commit/df400939addf6dbc5f2a9e1d52a6219f356f82d8 fixes
a bug that's prevalent on tota11y master, which causes JS errors when iframes of other origins are on the page.

Until that commit makes its way to npm, I am patching the method. I'm also trying a new pattern of explicitly
called a function named patch_____() - which should be easier to remove in the future. Perhaps this could even
go in a patches.js!

We should probably think about maintaining an ADT fork as well, but this works for now.

Test plan:
. npm run live-test
. Activate "Alt text" and "Labels" plugins
. See no JS errors

Auditors: michelle, riley