Create Plugin component and get LandmarksPlugin running
Summary:
This changeset introduces a `<Plugin>` component that renders an empty
`<PluginRunner>` to run and cleanup after its containing plugin. Also in place
are styles for the plugin controls to match tota11y.

{F210023}

Next steps are to start changing the Landmarks plugin to better work in this
style.

Test Plan:
* npm start
* localhost:1337/test
* Verify that the Landmarks plugin is listed
* Verify that you can toggle it on and off

Reviewers: michelle, riley

Reviewed By: michelle, riley

Differential Revision: https://phabricator.khanacademy.org/D23368