Convert LandmarksPlugin to use the new Plugin API
Summary:
This changeset modifies Landmarks plugin and the `<PluginRunner>` component to use the new plugin API. A few things are different, namely:

* Plugins no longer subclass anything
* `getTitle()` and `getDescription()` have been replaced with `title` and `description`
* Plugins no longer maintain their own `annotate` instance, but instead use an instance provided to them
* Plugins no longer need to clean up after themselves

You'll see the resulting Landmarks plugin is now much simpler. The next steps are to change the annotation logic to queue up animations and display them in a react way. This way we can rearchitect tota11y to run in the background and only //display// errors when necessary.

Test Plan:
* `npm start`
* http://localhost:8080/test
* See the Landmarks plugin successfully labels stuff and removes the labels when finished

Reviewers: riley, michelle

Reviewed By: michelle

Differential Revision: https://phabricator.khanacademy.org/D23569